chore(HttpClient): clean up catch regression coverage
